+++ src/parsing/pretty.ml
@@ -245,8 +245,8 @@ let query : p_query pp = fun ppf { elt; _ } ->
   match elt with
   | P_query_assert(true, a) -> out ppf "assertnot ⊢ %a" assertion a
   | P_query_assert(false,a) -> out ppf "assert ⊢ %a" assertion a
-  | P_query_debug(true ,s) -> out ppf "debug \"+%s\"" s
-  | P_query_debug(false,s) -> out ppf "debug \"-%s\"" s
+  | P_query_debug(true ,s) -> out ppf "debug +%s" s
+  | P_query_debug(false,s) -> out ppf "debug -%s" s
   | P_query_flag(s, b) ->
       out ppf "flag \"%s\" %s" s (if b then "on" else "off")
   | P_query_infer(t, _) -> out ppf "type %a" term t