+++ CHANGES.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+### Added
+
+- Incremental local confluence checking for non higher-order and non AC rules.
+
 ## 2.1.0 (2022-01-17)
 
 ### Added